fix(x64,client): client compiles clean at /W3 with 0 warnings (all proper fixes)
Completes the client x64 compile: 0 warnings, 0 compile errors at /W3, no
suppression anywhere. (Final link still pending x64 dep libs.)

- Fix the 687 C4244/C4996 sites surfaced by removing the suppressions, plus the
  cross-file deferrals: FrameController return type (BYTE->DWORD), Util __strcat1
  (manual append, no strcat), EterBase indexing helpers (size_t-bounded
  strcat_s/sprintf_s/_ecvt_s), GrpText atoi(_atoi64) narrowing made explicit.
- Client item-socket timestamp migration: TItemData.alSockets long->TimeT64 to
  match the 64-bit wire (packet.h). Narrow explicitly only at the client's stable
  DWORD metin-socket display contract (GetItemMetinSocket, item-link %x, the
  Python "i" socket accessors), preserving full width in storage.
- EterPythonLib: define HAVE_SNPRINTF so Python's pyerrors.h uses the standard
  snprintf instead of aliasing the deprecated _snprintf (matches sibling projects).
- MSB8012: align all 15 client static-lib x64 <Lib><OutputFile> to
  $(OutDir)$(TargetName)$(TargetExt) so TargetPath matches the library output.

+++ Lead-Client-Source/EffectLib/FrameController.cpp
@@ -41,7 +41,7 @@ void CFrameController::SetCurrentFrame(DWORD dwFrame)
 	m_dwcurFrame = dwFrame;
 }
 
-BYTE CFrameController::GetCurrentFrame()
+DWORD CFrameController::GetCurrentFrame()
 {
 	return m_dwcurFrame;
 }
