+++ Lead-Client-Source/ScriptLib/PythonMarshal.cpp
@@ -51,7 +51,7 @@ static int r_string(char *s, int n, RFILE *p)
 	if (p->fp != NULL)
 		return (int)fread(s, 1, n, p->fp); /* bounded by n (int); r_string contract returns int */
 	if (p->end - p->ptr < n)
-		n = p->end - p->ptr;
+		n = static_cast<int>(p->end - p->ptr); /* bounded: branch entered only when difference < n (int) */
 	memcpy(s, p->ptr, n);
 	p->ptr += n;
 	return n;
